backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
twilio>=9.3.0
apscheduler>=3.10.4
